Mongoose : implémentation du middleware d'authentification

+++ backend/controllers/sauces.js
@@ -12,9 +12,9 @@ exports.createSauce = (req, res, next) => {
 };
 
 exports.likeSauce = (req, res, next) => {
-    res.status(201).json({
-        message: ""
-    });
+    // res.status(201).json({
+    //     message: ""
+    // });
 };
 
 exports.modifySauce = (req, res, next) => {
@@ -24,8 +24,19 @@ exports.modifySauce = (req, res, next) => {
 };
 
 exports.deleteSauce = (req, res, next) => {
-    Sauce.deleteOne({_id: req.params.id})
-        .then(() => res.status(200).json({message: "Sauce supprimée."}))
+    Sauce.findOne({_id: req.params.id})
+        .then((sauce) => {
+            if (!sauce) {
+                return res.status(404).json({error: new Error("Sauce non trouvée.")});
+            }
+            // We check that the user making the request is the owner of the sauce
+            if (sauce.userId !== req.auth.userId) {
+                return res.status(401).json({error: new Error("Requête non autorisée.")});
+            }
+            Sauce.deleteOne({_id: req.params.id})
+                .then(() => res.status(200).json({message: "Sauce supprimée."}))
+                .catch((error) => res.status(400).json({error}));
+        })
         .catch((error) => res.status(400).json({error}));
 };
